Move part of models into data module
