+++ app/src/main/java/com/vultisig/wallet/data/mappers/KeysignPayloadProtoMapper.kt
@@ -5,14 +5,14 @@ import com.vultisig.wallet.data.api.models.OneInchSwapTxJson
 import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.Coin
 import com.vultisig.wallet.data.models.OneInchSwapPayloadJson
-import com.vultisig.wallet.data.models.SwapPayload
+import com.vultisig.wallet.data.models.THORChainSwapPayload
 import com.vultisig.wallet.data.models.payload.BlockChainSpecific
 import com.vultisig.wallet.data.models.payload.ERC20ApprovePayload
+import com.vultisig.wallet.data.models.payload.SwapPayload
 import com.vultisig.wallet.data.models.payload.UtxoInfo
 import com.vultisig.wallet.data.models.proto.v1.CoinProto
 import com.vultisig.wallet.data.models.proto.v1.KeysignPayloadProto
 import com.vultisig.wallet.data.models.proto.v1.ThorChainSwapPayloadProto
-import com.vultisig.wallet.models.THORChainSwapPayload
 import com.vultisig.wallet.presenter.keysign.KeysignPayload
 import java.math.BigDecimal
 import java.math.BigInteger
