+++ app/src/main/java/com/vultisig/wallet/data/models/SwapTransaction.kt
@@ -1,5 +1,6 @@
 package com.vultisig.wallet.data.models
 
+import com.vultisig.wallet.data.models.payload.SwapPayload
 import com.vultisig.wallet.data.repositories.BlockChainSpecificAndUtxo
 import java.math.BigInteger
