+++ app/src/main/java/com/vultisig/wallet/presenter/keysign/KeysignPayload.kt
@@ -13,10 +13,10 @@ import com.vultisig.wallet.chains.THORChainSwaps
 import com.vultisig.wallet.chains.UtxoHelper
 import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.Coin
-import com.vultisig.wallet.data.models.SwapPayload
 import com.vultisig.wallet.data.models.Vault
 import com.vultisig.wallet.data.models.payload.BlockChainSpecific
 import com.vultisig.wallet.data.models.payload.ERC20ApprovePayload
+import com.vultisig.wallet.data.models.payload.SwapPayload
 import com.vultisig.wallet.data.models.payload.UtxoInfo
 import com.vultisig.wallet.data.wallet.OneInchSwap
 import java.math.BigInteger
