+++ data/src/main/kotlin/com/vultisig/wallet/data/models/Account.kt
@@ -3,26 +3,26 @@ package com.vultisig.wallet.data.models
 import com.vultisig.wallet.data.models.settings.AppCurrency
 import java.math.BigDecimal
 
-internal data class Address(
+data class Address(
     val chain: Chain,
     val address: String,
     val accounts: List<Account>
 )
 
-internal data class Account(
+data class Account(
     val token: Coin,
     val tokenValue: TokenValue?,
     val fiatValue: FiatValue?
 )
 
-internal fun List<Account>.calculateAccountsTotalFiatValue(): FiatValue? =
+fun List<Account>.calculateAccountsTotalFiatValue(): FiatValue? =
     this.fold(FiatValue(BigDecimal.ZERO, AppCurrency.USD.ticker)) { acc, account ->
         // if any account dont have fiat value, return null, as in "not loaded yet"
         val fiatValue = account.fiatValue ?: return@calculateAccountsTotalFiatValue null
         FiatValue(acc.value + fiatValue.value, fiatValue.currency)
     }
 
-internal fun List<Address>.calculateAddressesTotalFiatValue(): FiatValue? =
+fun List<Address>.calculateAddressesTotalFiatValue(): FiatValue? =
     this.fold(FiatValue(BigDecimal.ZERO, AppCurrency.USD.ticker)) { acc, account ->
         // if any account dont have fiat value, return null, as in "not loaded yet"
         val fiatValue = account.accounts.calculateAccountsTotalFiatValue()
