+++ data/src/main/kotlin/com/vultisig/wallet/data/models/DepositTransaction.kt
@@ -2,7 +2,7 @@ package com.vultisig.wallet.data.models
 
 import com.vultisig.wallet.data.models.payload.BlockChainSpecific
 
-internal data class DepositTransaction(
+data class DepositTransaction(
     val id: TransactionId,
     val vaultId: String,
