+++ data/src/main/kotlin/com/vultisig/wallet/data/models/OneInchSwapPayloadJson.kt
@@ -4,7 +4,7 @@ import com.vultisig.wallet.data.api.models.OneInchSwapQuoteJson
 import java.math.BigDecimal
 import java.math.BigInteger
 
-internal data class OneInchSwapPayloadJson(
+data class OneInchSwapPayloadJson(
     val fromCoin: Coin,
     val toCoin: Coin,
     val fromAmount: BigInteger,
