+++ data/src/main/kotlin/com/vultisig/wallet/data/models/THORChainSwapPayload.kt
@@ -1,13 +1,11 @@
-package com.vultisig.wallet.models
+package com.vultisig.wallet.data.models
 
 import com.google.gson.annotations.SerializedName
-import com.vultisig.wallet.data.models.Chain
-import com.vultisig.wallet.data.models.Coin
 import wallet.core.jni.proto.THORChainSwap.Asset
 import java.math.BigDecimal
 import java.math.BigInteger
 
-internal data class THORChainSwapPayload(
+data class THORChainSwapPayload(
     @SerializedName("fromAddress")
     val fromAddress: String,
     @SerializedName("fromCoin")
